Refactor BrandLayout into layout subcomponents
Extract layout-related code from BrandLayout into dedicated files: BrandFooter, BrandHeaders (TopbarHeader/MinimalHeader), ControlsSection, DrawerNav, SidebarNav, UserSection, layout.constants, and layout.utils. BrandLayout now imports these modules and memoizes the drawer trigger with useMemo. Components are memoized/useCallback where appropriate and NAV_LINKS + helper utilities were centralized to improve modularity, readability, and render stability.
