+++ apps/ottabase-template-app-tanstack/src/ottabase/components/BrandLayout.tsx
@@ -1,321 +1,12 @@
-import { LanguageSwitcher } from '@/components/LanguageSwitcher';
-import { OrganizationSwitcher } from '@/components/OrganizationSwitcher';
 import { ReferralTracker } from '@/components/ReferralTracker';
-import { useLocalStorage } from '@/hooks/useLocalStorage';
-import { useSession } from '@/lib/auth';
-import { ThemeSwitcher } from '@/ottabase/components/ThemeSwitcher';
-import { APP_META } from '@/ottabase/config/app.config';
-import { i18nConfig } from '@/ottabase/config/i18n.config';
 import { useTheme } from '@/ottabase/providers/ThemeContext';
-import type { LayoutConfig } from '@ottabase/brand-engine';
-import { DarkModeToggle } from '@ottabase/ui-components/dark-mode-toggle';
-import { Avatar, AvatarFallback, AvatarImage, Button } from '@ottabase/ui-shadcn';
-import { Link, Outlet, useLocation, useNavigate } from '@tanstack/react-router';
-import { LogIn, LogOut, Menu, X } from 'lucide-react';
-import { useState } from 'react';
-import { createPortal } from 'react-dom';
-
-// ---------------------------------------------------------------------------
-// Nav link definitions (shared across all layout variants)
-// ---------------------------------------------------------------------------
-
-interface NavLink {
-    to: string;
-    label: string;
-    authRequired?: boolean;
-}
-
-const NAV_LINKS: NavLink[] = [
-    { to: '/', label: 'Home' },
-    { to: '/blog', label: 'Blog' },
-    { to: '/demo', label: 'Demo' },
-    { to: '/shortlinks', label: 'Links' },
-    { to: '/admin', label: 'Admin' },
-    { to: '/dashboard', label: 'Dashboard', authRequired: true },
-    { to: '/referrals', label: 'Referrals', authRequired: true },
-];
-
-// ---------------------------------------------------------------------------
-// Helpers
-// ---------------------------------------------------------------------------
-
-function contentWidthClass(contentWidth: LayoutConfig['contentWidth']): string {
-    switch (contentWidth) {
-        case 'fixed':
-            return 'max-w-5xl';
-        case 'fluid':
-            return 'max-w-7xl';
-        case 'full':
-            return 'w-full';
-        default:
-            return 'max-w-5xl';
-    }
-}
-
-function densityPadding(density: LayoutConfig['density']): string {
-    return density === 'compact' ? 'py-4' : 'py-10';
-}
-
-// ---------------------------------------------------------------------------
-// User info
-// ---------------------------------------------------------------------------
-
-function UserSection({ compact }: { compact?: boolean }) {
-    const { isAuthenticated, user, logout } = useSession();
-    const navigate = useNavigate();
-
-    const handleLogout = () => {
-        logout();
-        navigate({ to: '/' });
-    };
-
-    const userInitials =
-        user?.name && user.name.trim().length > 0
-            ? user.name
-                  .split(' ')
-                  .filter((n: string) => n.length > 0)
-                  .map((n: string) => n[0])
-                  .join('')
-                  .toUpperCase()
-            : user?.email && user.email.length > 0
-              ? user.email[0].toUpperCase()
-              : '?';
-
-    if (!isAuthenticated) {
-        return (
-            <div className={`flex items-center gap-2 ${compact ? '' : 'ml-2'}`}>
-                <Button asChild variant="ghost" size="sm">
-                    <Link to="/register">Sign up</Link>
-                </Button>
-                <Button asChild variant="default" size="sm">
-                    <Link to="/login" className="flex items-center gap-2">
-                        <LogIn className="h-4 w-4" />
-                        Login
-                    </Link>
-                </Button>
-            </div>
-        );
-    }
-
-    return (
-        <div className={`flex items-center gap-2 ${compact ? '' : 'ml-2 pl-2 border-l'}`}>
-            <Button asChild variant="ghost" size="sm">
-                <Link to="/profile" className="flex items-center gap-2">
-                    <Avatar className="h-6 w-6">
-                        {user?.image && <AvatarImage src={user.image} />}
-                        <AvatarFallback className="text-xs">{userInitials}</AvatarFallback>
-                    </Avatar>
-                    {!compact && (user?.name || user?.email)}
-                </Link>
-            </Button>
-            <Button variant="ghost" size="sm" onClick={handleLogout} title="Logout">
-                <LogOut className="h-4 w-4" />
-            </Button>
-        </div>
-    );
-}
-
-// ---------------------------------------------------------------------------
-// Controls (theme switcher, dark mode, language, org)
-// ---------------------------------------------------------------------------
-
-function ControlsSection() {
-    const { isAuthenticated } = useSession();
-    const [currentOrgId, setCurrentOrgId] = useLocalStorage<string>('currentOrgId');
-
-    return (
-        <div className="flex items-center gap-1">
-            <ThemeSwitcher />
-            <DarkModeToggle type="button" title="Toggle dark/light mode" />
-            <LanguageSwitcher languages={i18nConfig.enabledLanguages} showLabel={false} />
-            {isAuthenticated && <OrganizationSwitcher currentOrgId={currentOrgId} onOrgChange={setCurrentOrgId} />}
-        </div>
-    );
-}
-
-// ---------------------------------------------------------------------------
-// Drawer Navigation (mobile-first hamburger menu)
-// ---------------------------------------------------------------------------
-
-function DrawerNav() {
-    const { isAuthenticated } = useSession();
-    const location = useLocation();
-    const [open, setOpen] = useState(false);
-
-    return (
-        <>
-            <Button variant="ghost" size="sm" className="h-9 w-9 px-0" onClick={() => setOpen(true)} title="Open menu">
-                <Menu className="h-5 w-5" />
-            </Button>
-
-            {/* Portal to document.body so the overlay is not clipped by the
-                sticky header's backdrop-blur stacking context */}
-            {open &&
-                createPortal(
-                    <>
-                        <div
-                            className="fixed inset-0 bg-black/40 z-50"
-                            onClick={() => setOpen(false)}
-                            role="presentation"
-                        />
-                        <div
-                            className="fixed inset-y-0 left-0 w-64 bg-sidebar border-r z-50 flex flex-col animate-in slide-in-from-left duration-200"
-                            role="dialog"
-                            aria-modal="true"
-                            aria-label="Navigation Menu"
-                        >
-                            <div className="flex items-center justify-between p-4 border-b">
-                                <span className="font-semibold text-sm">{APP_META.appName}</span>
-                                <Button
-                                    variant="ghost"
-                                    size="sm"
-                                    className="h-8 w-8 px-0"
-                                    onClick={() => setOpen(false)}
-                                    aria-label="Close menu"
-                                >
-                                    <X className="h-4 w-4" />
-                                </Button>
-                            </div>
-                            <nav className="flex flex-col gap-0.5 p-3 overflow-y-auto flex-1">
-                                {NAV_LINKS.filter((l) => !l.authRequired || isAuthenticated).map((link) => {
-                                    const isActive =
-                                        location.pathname === link.to ||
-                                        (link.to !== '/' && location.pathname.startsWith(link.to));
-                                    return (
-                                        <Link
-                                            key={link.to}
-                                            to={link.to}
-                                            className={`px-3 py-2 text-sm rounded-md transition-colors ${
-                                                isActive
-                                                    ? 'bg-sidebar-accent text-sidebar-accent-foreground font-medium'
-                                                    : 'text-sidebar-foreground hover:bg-sidebar-accent/50'
-                                            }`}
-                                            onClick={() => setOpen(false)}
-                                        >
-                                            {link.label}
-                                        </Link>
-                                    );
-                                })}
-                            </nav>
-                        </div>
-                    </>,
-                    document.body,
-                )}
-        </>
-    );
-}
-
-// ---------------------------------------------------------------------------
-// Sidebar Navigation (desktop persistent sidebar)
-// ---------------------------------------------------------------------------
-
-function SidebarNav() {
-    const { isAuthenticated } = useSession();
-    const location = useLocation();
-    const links = NAV_LINKS.filter((l) => !l.authRequired || isAuthenticated);
-
-    return (
-        <aside className="w-full border-b bg-sidebar md:w-56 md:shrink-0 md:border-b-0 md:border-r md:sticky md:top-14 md:h-[calc(100dvh-3.5rem)] md:overflow-y-auto">
-            <nav className="flex gap-1 p-2 overflow-x-auto md:flex-col md:gap-0.5 md:p-3 md:overflow-x-visible">
-                {links.map((link) => {
-                    const isActive =
-                        location.pathname === link.to || (link.to !== '/' && location.pathname.startsWith(link.to));
-                    return (
-                        <Link
-                            key={link.to}
-                            to={link.to}
-                            className={`px-3 py-2 text-sm rounded-md transition-colors whitespace-nowrap md:whitespace-normal ${
-                                isActive
-                                    ? 'bg-sidebar-accent text-sidebar-accent-foreground font-medium'
-                                    : 'text-sidebar-foreground hover:bg-sidebar-accent/50'
-                            }`}
-                        >
-                            {link.label}
-                        </Link>
-                    );
-                })}
-            </nav>
-        </aside>
-    );
-}
-
-// ---------------------------------------------------------------------------
-// Header variants
-// ---------------------------------------------------------------------------
-
-function TopbarHeader({
-    showNav,
-    containerClass,
-    leading,
-}: {
-    showNav: boolean;
-    containerClass: string;
-    leading?: React.ReactNode;
-}) {
-    const { isAuthenticated } = useSession();
-
-    return (
-        <header className="border-b bg-background/95 backdrop-blur supports-[backdrop-filter]:bg-background/60 sticky top-0 z-40">
-            <div className={`mx-auto flex items-center justify-between px-4 py-3 ${containerClass}`}>
-                <div className="flex items-center gap-2">
-                    {leading}
-                    <Link to="/" className="font-semibold">
-                        {APP_META.appName}
-                    </Link>
-                    <span className="text-xs text-muted-foreground">TanStack</span>
-                </div>
-
-                {showNav && (
-                    <nav className="hidden md:flex items-center gap-1">
-                        {NAV_LINKS.filter((l) => !l.authRequired || isAuthenticated).map((link) => (
-                            <Button asChild variant="ghost" size="sm" key={link.to}>
-                                <Link to={link.to}>{link.label}</Link>
-                            </Button>
-                        ))}
-                    </nav>
-                )}
-
-                <div className="flex items-center gap-2">
-                    <ControlsSection />
-                    <UserSection />
-                </div>
-            </div>
-        </header>
-    );
-}
-
-function MinimalHeader({ containerClass, leading }: { containerClass: string; leading?: React.ReactNode }) {
-    return (
-        <header className="border-b bg-background/95 backdrop-blur supports-[backdrop-filter]:bg-background/60 sticky top-0 z-40">
-            <div className={`mx-auto flex items-center justify-between px-4 py-2 ${containerClass}`}>
-                <div className="flex items-center gap-2">
-                    {leading}
-                    <Link to="/" className="font-semibold text-sm">
-                        {APP_META.appName}
-                    </Link>
-                </div>
-                <div className="flex items-center gap-2">
-                    <ControlsSection />
-                    <UserSection compact />
-                </div>
-            </div>
-        </header>
-    );
-}
-
-// ---------------------------------------------------------------------------
-// Footer
-// ---------------------------------------------------------------------------
-
-function BrandFooter({ containerClass }: { containerClass: string }) {
-    return (
-        <footer className="border-t mt-auto">
-            <div className={`mx-auto px-4 py-6 text-center text-xs text-muted-foreground ${containerClass}`}>
-                Built with Ottabase
-            </div>
-        </footer>
-    );
-}
+import { Outlet } from '@tanstack/react-router';
+import { useMemo } from 'react';
+import { BrandFooter } from './layout/BrandFooter';
+import { MinimalHeader, TopbarHeader } from './layout/BrandHeaders';
+import { DrawerNav } from './layout/DrawerNav';
+import { SidebarNav } from './layout/SidebarNav';
+import { contentWidthClass, densityPadding } from './layout/layout.utils';
 
 // ---------------------------------------------------------------------------
 // BrandLayout – the root layout component driven by BrandEngine config
@@ -338,7 +29,8 @@ export function BrandLayout() {
     const navInHeader = navigation === 'topbar';
 
     // Drawer trigger shown in header for drawer navigation mode
-    const drawerTrigger = hasDrawer ? <DrawerNav /> : undefined;
+    // Memoized to maintain referential equality when passed to memoized headers
+    const drawerTrigger = useMemo(() => (hasDrawer ? <DrawerNav /> : undefined), [hasDrawer]);
 
     return (
         <div className="min-h-screen bg-background font-sans flex flex-col">
